Updated displayed information
This change adds more information to the autocompletion box, such as which file it was found in and the return type.
This also prevents deprecated functions from appearing within the autocompletion.

It has been reported that the previous PR may have broken something, although I'm not sure what. This version has been tested and works fine. Have been hesitant to create PR because I was hoping to finish working on new syntax methodmaps.
https://i.imgur.com/UJ717YH.png